requirements.txt
aiofiles==24.1.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-property==0.2.2
attrs==25.3.0
Authlib==1.5.2
bcrypt==4.0.1
cached-property==1.5.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.3
deprecation==2.1.0
dnspython==2.7.0
docstring_parser==0.16
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
Faker==37.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
FastAPI-SQLAlchemy==0.2.1
google-auth==2.40.1
greenlet==3.2.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.26.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
minio==7.2.15
mysql-connector-python==9.3.0
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pluggy==1.5.0
psutil==6.1.1
# psycopg2-binary==2.9.10  # PostgreSQL (commented - not using)
oracledb==2.5.1  # Oracle Cloud
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
pyodbc==5.2.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-editor==1.0.4
python-keycloak==5.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.5
rsa==4.9.1
setuptools==69.5.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
text-unidecode==1.3
toml==0.10.2
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wheel==0.43.0
zipp==3.21.0

# OCR & AI Services (from BackendNhoApp)
pytesseract==0.3.10
Pillow==10.1.0
aiohttp==3.9.1
paddlepaddle==3.0.0b1
paddleocr==2.9.1
lunarcalendar==0.0.9
openai==1.59.8

# HTTP Client for ASR microservice (using 0.26.0 - already defined above)
# httpx==0.28.1  # Duplicate - removed

# Speech-to-Text: MOVED to separate asr-service (port 8002)
# Backend communicates with ASR service via HTTP
# See: ../asr-service/ for ASR implementation
